requirements.txt
alembic
SQLAlchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
pydantic
pytest
httpx
requests
rq==1.14.0
redis==5.3.0
openai==1.8.0
jsonschema==4.19.0
